Перевод на MVVM
