+++ app/src/main/java/ru/netology/nmedia/Post.kt
@@ -1,0 +1,7 @@
+package ru.netology.nmedia
+
+data class Post(
+    var isLiked: Boolean = false,
+    var likesCount: Int = 9999,
+    var sharesCount: Int = 0
+)
